data-splitter/requirements.txt
boto
requests
Flask
pandas
SQLAlchemy
python-dotenv